+++ packages/wallets/walletconnect/src/walletconnect.ts
@@ -1,6 +1,10 @@
 import type { StdSignDoc } from '@cosmjs/amino';
 import type { TxBodyEncodeObject } from '@cosmjs/proto-signing';
-import type { DepositParam, TransferParams } from '@thorswap-lib/toolbox-cosmos';
+import type {
+  BaseCosmosToolboxType,
+  DepositParam,
+  TransferParams,
+} from '@thorswap-lib/toolbox-cosmos';
 import { ApiUrl, Chain, ChainId, WalletOption } from '@thorswap-lib/types';
 import type { WalletConnectModalSign } from '@walletconnect/modal-sign-html';
 import type { SessionTypes, SignClientTypes } from '@walletconnect/types';
@@ -405,9 +409,41 @@ const connectWalletconnect =
         stagenet,
       });
 
+      const getAccount = async (address: string) => {
+        const account = await (toolbox as unknown as BaseCosmosToolboxType).getAccount(address);
+        const [walletconnectAccount] = (await walletconnect?.client.request({
+          chainId: THORCHAIN_MAINNET_ID,
+          topic: session.topic,
+          request: {
+            method: DEFAULT_COSMOS_METHODS.COSMOS_GET_ACCOUNTS,
+            params: {},
+          },
+        })) as {
+          address: string;
+          algo: string;
+          pubkey: string;
+        }[];
+
+        return {
+          ...account,
+          address: walletconnectAccount.address,
+          pubkey: {
+            type: walletconnectAccount.algo,
+            value: walletconnectAccount.pubkey,
+          },
+        };
+      };
+
       addChain({
         chain,
-        walletMethods: { ...toolbox, getAddress: () => address },
+        walletMethods: {
+          ...toolbox,
+          getAddress: () => address,
+          getAccount:
+            chain === Chain.THORChain
+              ? getAccount
+              : (toolbox as unknown as BaseCosmosToolboxType).getAccount,
+        },
         wallet: { address, balance: [], walletType: WalletOption.WALLETCONNECT },
       });
       return;