+++ packages/swapkit/swapkit-entities/CHANGELOG.md
@@ -1,5 +1,11 @@
 []
 
+## 1.0.0-development.122
+
+### Patch Changes
+
+- [#335](https://github.com/thorswap/SwapKit/pull/335) [`68d493d`](https://github.com/thorswap/SwapKit/commit/68d493dde0c631fa6d39e4029083436d89a8a16d) Thanks [@0xepicode](https://github.com/0xepicode)! - Fix BSC min amount
+
 ## 1.0.0-development.121
 
 ### Patch Changes