+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @thorswap-lib/keplr
 
+## 1.0.0-development.61
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thorswap-lib/toolbox-cosmos@1.0.0-development.166
+
 ## 1.0.0-development.60
 
 ### Patch Changes