+++ packages/wallets/ledger/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @thorswap-lib/ledger
 
+## 1.0.0-development.266
+
+### Patch Changes
+
+- [#337](https://github.com/thorswap/SwapKit/pull/337) [`064ebe2`](https://github.com/thorswap/SwapKit/commit/064ebe29ed6a631058ad87c6a071bfa9e21ec325) Thanks [@towanTG](https://github.com/towanTG)! - Fixes Thorchain pubkey
+
+## 1.0.0-development.265
+
+### Patch Changes
+
+- Updated dependencies [[`68d493d`](https://github.com/thorswap/SwapKit/commit/68d493dde0c631fa6d39e4029083436d89a8a16d)]:
+  - @thorswap-lib/swapkit-entities@1.0.0-development.122
+  - @thorswap-lib/helpers@0.0.0-development.114
+  - @thorswap-lib/toolbox-cosmos@1.0.0-development.166
+  - @thorswap-lib/toolbox-evm@1.0.0-development.169
+  - @thorswap-lib/toolbox-utxo@1.0.0-development.148
+
 ## 1.0.0-development.264
 
 ### Patch Changes